OHRM5X-1145-fixed test cases issue
