+++ symfony/plugins/orangehrmMaintenancePlugin/AccessStrategy/FormatValue/FormatWithNationality.php
@@ -32,7 +32,11 @@ class FormatWithNationality implements ValueFormatter
      */
     public function getFormattedValue($entityValue): ?string
     {
-        return $this->getNationalityService()->getNationalityById($entityValue)->getName();
+        $result=$this->getNationalityService()->getNationalityById($entityValue);
+        if(!is_null($result)){
+            return $result->getName();
+        }
+        return null;
     }
 
     /**
