+++ symfony/plugins/orangehrmMaintenancePlugin/test/AccessStrategy/FormatValue/FormatWithPassportTypeTest.php
@@ -17,22 +17,23 @@
  * Boston, MA  02110-1301, USA
  */
 
-namespace OrangeHRM\Maintenance\test\accessStrategy\FormatValue;
+namespace OrangeHRM\Tests\Maintenance\AccessStrategy\FormatValue;
 
-use OrangeHRM\Maintenance\AccessStrategy\FormatValue\FormatWithMembershipId;
+use OrangeHRM\Maintenance\AccessStrategy\FormatValue\FormatWithPassportType;
 use OrangeHRM\Tests\Util\TestCase;
 
-class FormatWithMembershipIdTest extends TestCase
+class FormatWithPassportTypeTest extends TestCase
 {
-    private FormatWithMembershipId $formatWithMembershipId;
+    private FormatWithPassportType $formatWithPassportType;
 
     protected function setUp(): void
     {
-        $this->formatWithMembershipId = new FormatWithMembershipId();
+        $this->formatWithPassportType = new FormatWithPassportType();
     }
 
     public function testGetFormattedValue()
     {
-        $this->assertEquals('membership 1', $this->formatWithMembershipId->getFormattedValue(1));
+        $this->assertEquals('Passport', $this->formatWithPassportType->getFormattedValue(1));
+        $this->assertEquals('Visa', $this->formatWithPassportType->getFormattedValue(2));
     }
 }
